# public/openapi/apis.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "isrvd API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 服务"
    }
  ],
  "paths": {
...
    "/account/2fa/totp/disable": {
      "post": {
        "operationId": "post_account_2fa_totp_disable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.TOTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "禁用 TOTP 二次验证",
        "tags": [
          "account"
        ]
      }
    },
    "/account/2fa/totp/enable": {
      "post": {
        "operationId": "post_account_2fa_totp_enable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.TOTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "启用 TOTP 二次验证",
        "tags": [
          "account"
        ]
      }
    },
    "/account/login": {
      "post": {
        "operationId": "post_account_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "账号密码登录",
        "tags": [
          "account"
        ]
      }
    },
    "/account/member": {
      "post": {
        "operationId": "post_account_member",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.MemberUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "创建成员",
        "tags": [
          "account"
        ]
      }
    },
    "/account/member/{username}": {
      "delete": {
        "operationId": "delete_account_member_username",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "删除成员",
        "tags": [
          "account"
        ]
      },
      "put": {
        "operationId": "put_account_member_username",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.MemberUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "更新成员",
        "tags": [
          "account"
        ]
      }
    },
...
    "/account/oidc/exchange": {
      "post": {
        "operationId": "post_account_oidc_exchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.OIDCExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "交换 OIDC 登录码",
        "tags": [
          "account"
        ]
      }
    },
...
    "/account/passkey/credential/{credentialID}": {
      "delete": {
        "operationId": "delete_account_passkey_credential_credentialID",
        "parameters": [
          {
            "in": "path",
            "name": "credentialID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "删除 Passkey 凭证",
        "tags": [
          "account"
        ]
      },
      "put": {
        "operationId": "put_account_passkey_credential_credentialID",
        "parameters": [
          {
            "in": "path",
            "name": "credentialID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.accountPasskeyRenameCredential"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "重命名 Passkey 凭证",
        "tags": [
          "account"
        ]
      }
    },
...
    "/account/passkey/login/begin": {
      "post": {
        "operationId": "post_account_passkey_login_begin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.accountPasskeyLoginBegin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "开始 Passkey 登录",
        "tags": [
          "account"
        ]
      }
    },
    "/account/passkey/login/finish": {
      "post": {
        "operationId": "post_account_passkey_login_finish",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "完成 Passkey 登录",
        "tags": [
          "account"
        ]
      }
    },
    "/account/passkey/register/begin": {
      "post": {
        "operationId": "post_account_passkey_register_begin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.accountPasskeyRegisterBegin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "开始 Passkey 绑定",
        "tags": [
          "account"
        ]
      }
    },
    "/account/passkey/register/finish": {
      "post": {
        "operationId": "post_account_passkey_register_finish",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "完成 Passkey 绑定",
        "tags": [
          "account"
        ]
      }
    },
    "/account/password": {
      "put": {
        "operationId": "put_account_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "修改当前用户密码",
        "tags": [
          "account"
        ]
      }
    },
...
    "/account/token": {
      "post": {
        "operationId": "post_account_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/account.CreateApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "创建 API 令牌",
        "tags": [
          "account"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {},
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
      "account.ChangePasswordRequest": {
        "properties": {
          "newPassword": {
            "description": "新密码",
            "type": "string"
          },
          "oldPassword": {
            "description": "原密码",
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ],
        "type": "object"
      },
      "account.CreateApiTokenRequest": {
        "properties": {
          "expiresIn": {
            "description": "过期时间（秒），0 表示永不过期",
            "type": "integer"
          },
          "name": {
            "description": "令牌名称（用于标识）",
            "type": "string"
          }
        },
        "type": "object"
      },
      "account.LoginRequest": {
        "properties": {
          "password": {
            "description": "密码",
            "type": "string"
          },
          "totpCode": {
            "description": "TOTP 验证码（启用二步验证时必填）",
            "type": "string"
          },
          "username": {
            "description": "用户名",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "account.MemberUpsertRequest": {
        "properties": {
          "description": {
            "description": "成员描述",
            "type": "string"
          },
          "homeDirectory": {
            "description": "家目录（绝对路径或基于 RootDirectory 的相对路径）",
            "type": "string"
          },
          "password": {
            "description": "密码（创建时必填，更新时为空则保留原密码）",
            "type": "string"
          },
          "permissions": {
            "description": "权限列表",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "username": {
            "description": "用户名（创建时必填，更新时从 URL 读取）",
            "type": "string"
          }
        },
        "type": "object"
      },
      "account.OIDCExchangeRequest": {
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "account.TOTPVerifyRequest": {
        "properties": {
          "code": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "account.accountPasskeyLoginBegin": {
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "account.accountPasskeyRegisterBegin": {
        "properties": {
          "displayName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "account.accountPasskeyRenameCredential": {
        "properties": {
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}